backend/requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
astor==0.7.1
astroid==2.2.5
astunparse==1.6.3
attrs==19.1.0
beautifulsoup4==4.7.1
blis==0.4.1
boto==2.49.0
boto3==1.14.38
botocore==1.17.38
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2019.3.9
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
cycler==0.10.0
cymem==2.0.2
decorator==4.4.1
Django==2.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
docutils==0.15.2
elasticsearch==7.9.1
en-core-web-sm==2.1.0
entrypoints==0.3
fire==0.3.1
flake8==3.7.7
Flask==1.0.2
Flask-RESTful==0.3.7
future==0.17.1
gast==0.3.3
gensim==3.8.3
google==2.0.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.20.1
gym==0.12.5
h5py==2.10.0
idna==2.8
importlib==1.0.4
importlib-metadata==1.7.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.13.2
jsonschema==3.0.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
logger==1.4
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.3.0
mccabe==0.6.1
mock==3.0.3
murmurhash==1.0.2
networkx==2.4
nlplib==0.1
nltk==3.4.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.0
pep8==1.7.1
Pillow==6.0.0
Pint==0.9
pipenv==2020.11.15
plac==0.9.6
plotly==4.9.0
preshed==3.0.2
protobuf==3.13.0
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyglet==1.3.2
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.0
PyPDF2==1.26.0
pyrsistent==0.15.2
python-dateutil==2.7.5
pytz==2019.1
PyYAML==3.13
regex==2020.9.27
requests==2.21.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.9.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
smart-open==2.1.0
soupsieve==1.9.1
spacy==2.3.2
SQLAlchemy==1.2.19
sqlparse==0.3.0
srsly==1.0.2
stop-words==2018.7.23
tabulate==0.8.3
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
termcolor==1.1.0
tesseract==0.1.3
textblob==0.15.3
texthero==1.0.9
tflearn==0.3.2
thinc==7.4.1
threadpoolctl==2.1.0
torch==1.5.0
tqdm==4.32.1
typed-ast==1.3.5
Unidecode==1.1.1
urllib3==1.24.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
wasabi==0.7.1
Werkzeug==0.15.2
wordcloud==1.7.0
wrapt==1.11.1
xlrd==1.2.0
zipp==3.1.0
